Lots of progress
I can click between x and o with x being first. Start New Game is up
and running with its own api and everything...not sure if that is
going to be the best way to go about it.
still working on game logic. and letting the player know what they
play.